ScoreDisplay % ClickerButton scripts cooperating
ScoreDisplay now behaves as intended with ClickerButton.

+++ Assets/Scripts/QuestUnlock.cs
@@ -6,7 +6,7 @@ public class QuestUnlock : MonoBehaviour
 {
     [SerializeField] private ClickerButton button;
 
-    int scoreThreshold = 0;
+    float scoreThreshold = 0;
 
     [SerializeField] GameObject Quest1;
     [SerializeField] GameObject Quest2;
